+++ stories/Ellipsis/Ellipsis.stories.tsx
@@ -1,0 +1,68 @@
+import React from "react";
+import { Meta, StoryObj } from "@storybook/react";
+import { Title, Description, Primary, Controls, Stories, Heading, ArgTypes } from "@storybook/addon-docs/blocks";
+import { Ellipsis } from "@sberbusiness/triplex-next";
+import { DefaultExample, DefaultExampleSource } from "./examples/index";
+
+const meta = {
+    title: "Components/Ellipsis",
+    component: Ellipsis,
+    tags: ["autodocs"],
+    parameters: {
+        docs: {
+            description: {
+                component: `
+Компонент для сворачивания в многоточие текста, который не поместился в заданное количество строк. Данному компоненту нельзя устанавливать паддинги, так как реализация через CSS свойство line-clamp.
+                `,
+            },
+            page: () => (
+                <>
+                    <Title />
+                    <Description />
+                    <Heading>Props</Heading>
+                    <ArgTypes of={Ellipsis} />
+                    <Heading>Playground</Heading>
+                    <Primary />
+                    <Controls of={Playground} />
+                    <Stories />
+                </>
+            ),
+        },
+    },
+} satisfies Meta<typeof Ellipsis>;
+
+export default meta;
+type Story = StoryObj<typeof Ellipsis>;
+
+export const Playground: Story = {
+    tags: ["!autodocs"],
+    args: { maxLines: 2 },
+    argTypes: {
+        maxLines: {
+            control: { type: "number" },
+            description: "Количество строк, после которых происходит сворачивание в многоточие.",
+        },
+    },
+    parameters: {
+        controls: { include: ["maxLines"] },
+        docs: { canvas: { sourceState: "none" }, codePanel: false },
+        testRunner: { skip: true },
+    },
+    render: (args) => (
+        <Ellipsis {...args}>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et
+            dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex
+            ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat
+            nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
+            anim id est laborum.
+        </Ellipsis>
+    ),
+};
+
+export const Default: Story = {
+    render: DefaultExample,
+    parameters: {
+        controls: { disable: true },
+        docs: { source: { code: DefaultExampleSource, language: "tsx" } },
+    },
+};
