+++ stories/Ellipsis/examples/index.ts
@@ -1,0 +1,2 @@
+export { DefaultExample } from "./DefaultExample";
+export { default as DefaultExampleSource } from "./DefaultExample?raw";
